+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 ## 0.41.0 - TBD
 
 ### Enhancements
-- Add static `Builder()` methods to the clients
-- Improve debug logging in live clients
+- Added static `Builder()` methods to the clients
+- Improved debug logging in live clients
+- Added `PUBLISHER_SPECIFIC` flag
 
 ### Breaking changes
 - Removed unused `Received` variant from `JobState` enum